first complete draft
I'm a total beginner so hopefully this will help someone else in figuring out some FastLED features. I'm fairly certain I still have editing to do as there was a lot of trial and error. Special thanks to Mark Kriegsman and the FastLED Google+ group for leading me in the right direction!